third_party/blink/web_tests/media/controls/download-button-displays-with-preload-none.html

<!DOCTYPE html>
<title>media controls download button preload none</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/run-after-layout-and-paint.js"></script>
<script src="../media-controls.js"></script>
<video controls preload="none" src="https://someexample.example/example.mp4"></video>
<script>
async_test(function(t) {
  var video = document.querySelector("video");

  runAfterLayoutAndPaint(t.step_func_done(function() {
    assert_true(isDownloadsButtonEnabled(video));
  }));

});
</script>
